Reap orphan zombies
On Unix systems, PID 1 processes become the parent of orphan processes
(processes that out live their parents). This is generally `init`,
which will `wait(2)` on these orphan processes when they end in order
to free the resources consumed by them in the kernel's process table
("reaping"). If a process is not waited on, it is a _zombie_; if
enough zombies accumulate on a system, the `nproc` limit can be
exceeded, making it impossible to start new processes.

When the Habitat Launcher is running as PID 1 (as is the case in a
Habitat-exported Docker container), it needs to reap orphan processes
that it adopts by virtue of being PID 1. We already do this for
Habitat service processes that are being monitored, but this applies
to any other processes that can be spun up in the process of running a
service (the `core/rabbitmq`, for instance, service currently generates a number of
orphans from its `health_check` hook; see habitat-sh/core-plans#1099).

This commit takes advantage of a `SIGCHLD` handler to notify the
Launcher whenever any of its children end, allowing it to properly
reap them.

"Children of the Launcher" (in addition to making a cool dystopian
sci-fi movie name) also includes the Supervisor process itself. We
take care to check if a reaped child process is our Supervisor and
react accordingly by restarting it or shutting it down. One change
from the previous implementation is that we now explicitly check to
see if the Supervisor was terminated by a signal. Previously, this
would ultimately result in the Launcher itself shutting down, but
through a non-obvious chain of Supervisor process restarts. To make
the behavior more obvious, we now explicitly shut down if the
Supervisor ends as a result of signal termination.

Note that we are not currently using `prctl(PR_SET_CHILD_SUBREAPER,...)`
on the Launcher, though that would be a future optimization to
make. As such, child processes are only reparented to the Launcher if
it is running as PID 1. Setting the Launcher to be a subreaper would
allow the Launcher to reap these children on systems where it is not
running as PID 1.

Since this is a Unix behavior, the Windows implementation is basically
a no-op.

Signed-off-by: Christopher Maier <cmaier@chef.io>